AlertSystem/requirements.txt
mysql-connector-python
confluent-kafka==1.9.2